requirements.txt
apache-airflow==2.9.0; python_version < "3.13"
pandas
scikit-learn
xgboost
mlflow
dvc
fastapi
uvicorn
pydantic
boto3
docker
prometheus-client
pyyaml
joblib
numpy
seaborn
matplotlib
